Organised classes into java packages
