+++ data/task_data.txt
@@ -1,5 +1,4 @@
 D | 1 | mario kart | 15/02/2024 1900
 T | 1 | lecture watching
 E | 0 | cs2103t class | 15/02/2024 1700 | 15/02/2024 1800
-E | 1 | play mario kart | 15/02/2024 1800 | 16/02/2024 1000
-E | 0 | hey | 15/02/2024 1800 | 16/02/2024 1000
+E | 0 | play mario kart | 15/02/2024 1800 | 16/02/2024 1000
