--- src/main/java/GreetCommand.java
@@ -1,9 +1,0 @@
-public class GreetCommand extends Command {
-
-    public GreetCommand() {
-        super();
-    }
-    public void execute(TaskList tasks, UI ui, Storage storage) {
-        ui.greet();
-    }
-}
