--- src/main/java/GuideCommand.java
@@ -1,6 +1,0 @@
-public class GuideCommand extends Command{
-    @Override
-    public void execute(TaskList tasks, UI ui, Storage storage) {
-        ui.printTutorial();
-    }
-}
