--- src/main/java/LogCommand.java
@@ -1,6 +1,0 @@
-public class LogCommand extends Command {
-    @Override
-    public void execute(TaskList tasks, UI ui, Storage storage) {
-        ui.printLog(tasks);
-    }
-}
