+++ src/main/java/bmo/command/Command.java
@@ -1,3 +1,9 @@
+package bmo.command;
+
+import bmo.util.Storage;
+import bmo.util.TaskList;
+import bmo.ui.Ui;
+
 public abstract class Command {
 
      private final boolean isExit;
@@ -5,7 +11,7 @@ public abstract class Command {
      public Command() {
           this.isExit = false;
      }
-     public abstract void execute(TaskList tasks, UI ui, Storage storage);
+     public abstract void execute(TaskList tasks, Ui ui, Storage storage);
 
      public boolean isExit() {
           return this.isExit;
