+++ src/main/java/bmo/command/DeleteCommand.java
@@ -1,3 +1,9 @@
+package bmo.command;
+
+import bmo.util.Storage;
+import bmo.util.TaskList;
+import bmo.ui.Ui;
+
 public class DeleteCommand extends Command{
 
     private final String index;
@@ -7,7 +13,7 @@ public class DeleteCommand extends Command{
     }
 
     @Override
-    public void execute(TaskList tasks, UI ui, Storage storage) {
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
         if (this.index.isBlank() || !this.index.matches("\\d+")) {
             ui.printErrInvalidIndex();
             return;
