+++ src/main/java/bmo/command/DoneCommand.java
@@ -1,3 +1,10 @@
+package bmo.command;
+
+import bmo.util.Storage;
+import bmo.util.TaskList;
+import bmo.ui.Ui;
+import bmo.task.Task;
+
 public class DoneCommand extends Command {
 
     private final String index;
@@ -7,7 +14,7 @@ public class DoneCommand extends Command {
     }
 
     @Override
-    public void execute(TaskList tasks, UI ui, Storage storage) {
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
         if (this.index.isBlank() || !this.index.matches("\\d+")) {
             ui.printErrInvalidIndex();
             return;
