+++ src/main/java/bmo/command/ExitCommand.java
@@ -1,6 +1,12 @@
+package bmo.command;
+
+import bmo.util.Storage;
+import bmo.util.TaskList;
+import bmo.ui.Ui;
+
 public class ExitCommand extends Command{
     @Override
-    public void execute(TaskList tasks, UI ui, Storage storage) {
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
         ui.salute();
         storage.saveData(tasks);
     }
