+++ src/main/java/bmo/command/FindCommand.java
@@ -1,3 +1,9 @@
+package bmo.command;
+
+import bmo.util.Storage;
+import bmo.util.TaskList;
+import bmo.ui.Ui;
+
 public class FindCommand extends Command {
     private String keyword;
 
@@ -6,7 +12,7 @@ public class FindCommand extends Command {
     }
 
     @Override
-    public void execute(TaskList tasks, UI ui, Storage storage) {
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
         System.out.println("BMO has not updated this feature yet");
     }
 }
