+++ src/main/java/bmo/command/GuideCommand.java
@@ -1,0 +1,12 @@
+package bmo.command;
+
+import bmo.util.Storage;
+import bmo.util.TaskList;
+import bmo.ui.Ui;
+
+public class GuideCommand extends Command{
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
+        ui.printTutorial();
+    }
+}
