+++ src/main/java/bmo/command/LogCommand.java
@@ -1,0 +1,12 @@
+package bmo.command;
+
+import bmo.util.Storage;
+import bmo.util.TaskList;
+import bmo.ui.Ui;
+
+public class LogCommand extends Command {
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
+        ui.printLog(tasks);
+    }
+}
