+++ src/main/java/bmo/command/RedoCommand.java
@@ -1,3 +1,10 @@
+package bmo.command;
+
+import bmo.util.Storage;
+import bmo.util.TaskList;
+import bmo.ui.Ui;
+import bmo.task.Task;
+
 public class RedoCommand extends Command {
 
     private final String index;
@@ -7,7 +14,7 @@ public class RedoCommand extends Command {
     }
 
     @Override
-    public void execute(TaskList tasks, UI ui, Storage storage) {
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
         if (this.index.isBlank() || !this.index.matches("\\d+")) {
             ui.printErrInvalidIndex();
             return;
