+++ src/main/java/bmo/command/ToDoCommand.java
@@ -1,5 +1,12 @@
+package bmo.command;
+
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
+import bmo.util.Storage;
+import bmo.util.TaskList;
+import bmo.ui.Ui;
+import bmo.task.Task;
+import bmo.task.ToDo;
 
 public class ToDoCommand extends Command{
 
@@ -8,7 +15,7 @@ public class ToDoCommand extends Command{
         this.desc = desc;
     }
 
-    public void execute(TaskList tasks, UI ui, Storage storage) {
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
         Task newTask = new ToDo(this.desc);
         tasks.add(newTask);
         ui.printAddTask(newTask, tasks.size());
