+++ src/main/java/bmo/task/Events.java
@@ -1,3 +1,5 @@
+package bmo.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
