+++ src/main/java/bmo/task/ToDo.java
@@ -1,3 +1,5 @@
+package bmo.task;
+
 public class ToDo extends Task {
 
     public ToDo(String task) {
