+++ src/main/java/bmo/util/Storage.java
@@ -1,8 +1,11 @@
+package bmo.util;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.util.Scanner;
+import bmo.task.Task;
 
 public class Storage {
